Add files via upload
The above file contains assignment where I had used images in html instead of css due to a system error. My css file in vs code was facing issues when connecting to an image file, so it was not possible to make the web page responsive or work with images with any freedom.